requirements.txt
numpy
scipy
torch
plotly
adult-dataset
